Adds configuration of conversion rates and RSpec
Why:

* Needed to configure the conversion rates to be used on the gem

This change addresses the need by:

* Storing the conversion rates in memory. This should be good enough for
a first version of the gem, even though integrating with Redis would be
a good next step. In a rails version of the gem, we could use the DB to
persist this conversion rates.
* Exposes methods to get a specific conversion rate, and to clear all
conversion rates.
* By using required kwargs we are assuming that we depend on Ruby 2.2 or
newer.
* Decided to go with RSpec instead of Minitest, because it is the most
used testing library and will probably be the one that more developers
will be used to.
* Runs the spec as the Rake default task.